Updated scope service
