+++ packages/server-core/src/user/identity-provider/identity-provider.class.ts
@@ -196,7 +196,7 @@ export class IdentityProviderService<
 
     const code = await getFreeInviteCode(this.app)
     // if there is no user with userId, then we create a user and a identity provider.
-    const adminCount = (await this.app.service(scopePath)._find({
+    const adminCount = (await this.app.service(scopePath).find({
       query: {
         type: 'admin:admin'
       }
